Removed Websockify dependency / Bugfixes
Was able to fix the "Rocket spasms out" issue, by not sending the row on
deletions (or ctrl+z) - and not setting the row in render when the demo
is paused anyways.
Also fixed an issue, that floats in Rocket (eg 400.10) have been parsed
as int (eg 400) - both have the same "unprecision".